requirements.txt
streamlit
pandas
plotly
python-dateutil